presentation/src/main/java/com/open/presentation/viewmodel/WeatherViewModel.kt
package com.open.presentation.viewmodel

import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.open.data.network.WeatherState
import com.open.domain.usecases.GetWeatherUseCase
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class WeatherViewModel @Inject constructor(private val getWeatherUseCase: GetWeatherUseCase) : ViewModel() {
    private val _weatherState = MutableStateFlow<WeatherState>(WeatherState.Loading)
    val weatherState: StateFlow<WeatherState> = _weatherState

    fun getWeather(city: String) {
        viewModelScope.launch {
            getWeatherUseCase(city).collect { response ->
                _weatherState.value = WeatherState.Success(response)
            }
        }
    }
}
